+++ src/App/Controller/AddUser.php
@@ -11,21 +11,29 @@ class AddUser extends AbstractController
 {
     public function __invoke(): string
     {
-        if (isset($_SESSION))
+        if (isset($_SESSION["user"]))
             header('Location: /');
         try {
-            $userCont = new UsersController();
-            $user = user::create($_GET);
-           // var_dump($user);die;
-            $userCont->addUser($user);
 
+            $content = trim(file_get_contents("php://input"));
+
+            $data = json_decode($content, true);
+            $data['success'] = true;
+
+
+
+
+            if ($data['success'])
+                $data = $data['jsonData'];
+
+            $userCont = new UsersController();
+            $user = user::create($data);
+            $userCont->AddUser($user);
+            return true; 
         } catch (Exception $e) {
-            return $this->render('error.html.twig', [
-                'error' => $e->getMessage(),
-            ]);
+            return  $e->getMessage();
         }
 
-
-        // header('Location: /connexion');
+ 
     }
 }